Refactor: Modularize AntiTilt equipment module
- Split apts/opticalequipment/anti_tilt.py into a package.
- Extract core logic to apts/opticalequipment/anti_tilt/base.py.
- Partition 35 hardware specs into 18 vendor modules in vendors/ subdirectory.
- Implement dynamic discovery/aggregation in __init__.py.
- Maintain public API for backward compatibility.
